Added in ADC Interrupt
Interrupt fires if I comment out the LED code, but only fires a number
of times with the LED code in there. Will have to troubleshoot. Also
need to start setting the ADC read in values inside the ADC interrupt
rather than in the ISR how it is currently. Put a global in the ADC to
make sure it’s working correctly.